Fix progress indicator CPU usage
This is a bit of a hack, but seems there is not way to set the fps of
ClutterTimeline, so new-frame cb is called very fast. Redrawing the
indicator on every new frame doesn't make much sense, so check if we really
need to redraw it.

Also, fix some remaining compiler warnings while at it.

Signed-off-by: Ivaylo Dimitrov <ivo.g.dimitrov.75@gmail.com>