configure: Remove HAVE_XFIXES
The Cordia patch to show/hide the mouse cursor if a (non-)touchscreen
input device is used didn't work, since libmatchbox2 doesn't use XFixes
to show/hide the cursor, but it just defines an empty cursor. This means
that even if we show the cursor, it will still not show up.

Using XUndefineCursor (and XDefineCursor) does work, since calling
XUndefineCursor on the root window loads the default cursor.